Be consistent in the values being used for the woosimple-toggle-switch
All of the logic has been counting on a boolean (with the meta value itself being stored as either a '1' or '0'), while some of the recent updates used "on" to indicate a truthy-state. However, type-juggling within PHP would interpret both "on" and "off" as truth-y, causing the value to be evaluated in unexpected ways.

Fixes #5.